Refactor the script generating the experimental design.
