+++ scripts/figures/spiral/generate_tasks.py
